requirements.txt
vk_api>=13.7
python-dotenv>=1.0.0
openai>=1.0.0
